Reorganize javascript to separate concerns and clean up. First of several passes.
