+++ dlx_rest/static/js/browse_vm.js
@@ -3,7 +3,7 @@
 /////////////////////////////////////////////////////////////////
 
 //import { Jmarc } from "./jmarc.mjs";
-import { browsecomponent } from "./browse.js";
+import { browsecomponent } from "./components/browse.js";
 
 /////////////////////////////////////////////////////////////////
 // VIEW MODEL DEFINITION
