--- dlx_rest/static/js/charts.js
@@ -1,70 +1,0 @@
-// Wait for both DOM and MongoDB Charts SDK to be ready
-function initializeCharts() {
-  try {
-    console.log('Initializing charts...');
-    console.log('DOM ready state:', document.readyState);
-    console.log('MongoDB Charts SDK available:', typeof window.ChartsEmbedSDK !== 'undefined');
-    console.log('userCheckboxes element:', document.getElementById('userCheckboxes'));
-
-    if (typeof window.ChartsEmbedSDK === 'undefined') {
-      console.log('Waiting for MongoDB Charts SDK to load...');
-      setTimeout(initializeCharts, 100);
-      return;
-    }
-
-    const ChartsEmbedSDK = window.ChartsEmbedSDK;
-    const sdk = new ChartsEmbedSDK({
-      baseUrl: "https://charts.mongodb.com/charts-project-0-vbkva",  
-    });
-
-    const dashboardId = '2c30ae76-5b8e-4703-af90-797962da36fd';
-    const userCheckboxesContainer = document.getElementById('userCheckboxes');
-    
-    if (!userCheckboxesContainer) {
-      console.error('Could not find userCheckboxes container. Make sure you are on the correct page.');
-      console.log('Current URL:', window.location.href);
-      console.log('Available elements:', document.body.innerHTML);
-      return;
-    }
-
-    // Create a function to update filters based on selected checkboxes
-    function updateDashboard() {
-      const selectedUsers = Array.from(document.querySelectorAll('input[type=checkbox]:checked'))
-        .map(checkbox => checkbox.value);
-
-      console.log("Selected users:", selectedUsers);
-     
-      // Create or update the dashboard with the selected user filters
-      const chart = sdk.createDashboard({
-        dashboardId: dashboardId,
-        filter: { user: { $in: selectedUsers} }            
-      });  
-      chart.render(document.getElementById("dashboard"));   
-    } 
- 
-    // Add an event listener to checkboxes to trigger updates
-    userCheckboxesContainer.addEventListener('change', function() {
-      updateDashboard();
-    });  
-  
-    // Check all checkboxes by default
-    document.querySelectorAll('input[type=checkbox]').forEach(checkbox => {
-      checkbox.checked = true;
-    });
- 
-    // Initial update with all users selected
-    updateDashboard();
-  } catch (error) {
-    console.error('Error initializing charts:', error);
-  }
-}
-
-// Start initialization when DOM is ready
-if (document.readyState === 'loading') {
-  console.log('DOM is still loading, waiting for DOMContentLoaded event...');
-  document.addEventListener('DOMContentLoaded', initializeCharts);
-} else {
-  console.log('DOM is already loaded, initializing immediately...');
-  initializeCharts();
-}
-
