+++ dlx_rest/static/js/components/agenda.js
