+++ dlx_rest/static/js/components/auth_review.js
@@ -1,7 +1,7 @@
-import { sortcomponent } from "./search/sort.js";
-import { countcomponent } from "./search/count.js";
-import basket from "./api/basket.js";
-import user from "./api/user.js";
+import { sortcomponent } from "./sort.js";
+import { countcomponent } from "./count.js";
+import basket from "../api/basket.js";
+import user from "../api/user.js";
 import { readonlyrecord } from "./readonly_record.js";
 
 export let authreviewcomponent = {
