+++ dlx_rest/static/js/components/count.js
