+++ dlx_rest/static/js/components/header.js
