+++ dlx_rest/static/js/components/itemadd.js
