+++ dlx_rest/static/js/components/merge.js
