+++ dlx_rest/static/js/components/messagebar.js
