+++ dlx_rest/static/js/components/preview.js
@@ -1,4 +1,4 @@
-import { Jmarc } from "../jmarc.mjs"
+import { Jmarc } from "../api/jmarc.mjs"
 
 export let previewmodal = {
     props: {
