+++ dlx_rest/static/js/components/recordfiles.js
