+++ dlx_rest/static/js/components/search.js
@@ -1,11 +1,11 @@
-import { sortcomponent } from "./search/sort.js";
-import { countcomponent } from "./search/count.js";
-import basket from "./api/basket.js";
-import user from "./api/user.js";
+import { sortcomponent } from "./sort.js";
+import { countcomponent } from "./count.js";
+import basket from "../api/basket.js";
+import user from "../api/user.js";
 import { readonlyrecord } from "./readonly_record.js"
 import { recordfilecomponent } from "./recordfiles.js";
-import { exportmodal } from "./modals/export.js";
-import { searchHistoryComponent } from "./components/search-history.js";
+import { exportmodal } from "./export.js";
+import { searchHistoryComponent } from "./search-history.js";
 
 export let searchcomponent = {
     props: {
