+++ dlx_rest/static/js/components/sidebar.js
