+++ dlx_rest/static/js/lib/vue2.js
