+++ dlx_rest/static/js/search_vm.js
@@ -3,17 +3,16 @@
 /////////////////////////////////////////////////////////////////
 
 //import { Jmarc } from "./jmarc.mjs";
-import { messagecomponent } from "./messagebar.js";
-import { sidebarcomponent } from "./search/sidebar.js";
-import { searchcomponent } from "./search.js";
-import { basketcomponent } from "./basket.js";
+import { messagecomponent } from "./components/messagebar.js";
+import { searchcomponent } from "./components/search.js";
+import { basketcomponent } from "./components/basket.js";
 
 /////////////////////////////////////////////////////////////////
 // VIEW MODEL DEFINITION
 /////////////////////////////////////////////////////////////////
 export let vm_search_component = new Vue({
     el: '#search_vm',
-    components: { sidebarcomponent, searchcomponent, basketcomponent },
+    components: { searchcomponent, basketcomponent },
     data: {
       visible: false,
     },
