+++ dlx_rest/static/js/utils/charts-init.js
